Add files via upload
For this project, the primary goal is to adopt a graph-based machine learning (ML) model to classify the licit and illicit Bitcoin transactions. To achieve this, both supervised and semi- supervised machine learning algorithms are utilized. Once the model with the best performance (e.g., high F1 score or accuracy) is established, it will be used to classify the unlabelled data points or records (e.g., Bitcoin transactions). All the models are built with Pyspark in Data Bricks Community edition.